ecom/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==25.1
mysqlclient==2.2.4
Pillow==10.4.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3